refactor: refactor sheets callbacks
